KeyboardCowboys/src/test/java/bcit/ca/infosys/KeyboardCowboys/test/integration/EmployeeRegistrationTest.java
package bcit.ca.infosys.KeyboardCowboys.test.integration;

import static org.junit.Assert.assertNotNull;

import java.io.File;
import java.util.ArrayList;
import java.util.List;
import java.util.logging.Logger;

import javax.inject.Inject;

import org.jboss.arquillian.container.test.api.Deployment;
import org.jboss.arquillian.junit.Arquillian;
import org.jboss.shrinkwrap.api.Archive;
import org.jboss.shrinkwrap.api.ShrinkWrap;
import org.jboss.shrinkwrap.api.asset.EmptyAsset;
import org.jboss.shrinkwrap.api.spec.WebArchive;
import org.jboss.shrinkwrap.resolver.api.maven.Maven;
import org.junit.Test;
import org.junit.runner.RunWith;

import bcit.ca.infosys.KeyboardCowboys.exceptions.EmployeeAlreadyExistsException;
import bcit.ca.infosys.KeyboardCowboys.interfaces.EmployeeRegistrationInterface;
import bcit.ca.infosys.KeyboardCowboys.model.Employee;
import bcit.ca.infosys.KeyboardCowboys.model.EmployeePayInfo;
import bcit.ca.infosys.KeyboardCowboys.model.PayLevel;
import bcit.ca.infosys.KeyboardCowboys.model.Project;
import bcit.ca.infosys.KeyboardCowboys.model.TimeRow;
import bcit.ca.infosys.KeyboardCowboys.model.TimeSheet;
import bcit.ca.infosys.KeyboardCowboys.model.TimeSheetApproval;
import bcit.ca.infosys.KeyboardCowboys.model.WorkPackage;
import bcit.ca.infosys.KeyboardCowboys.service.EmployeeRegistration;
import bcit.ca.infosys.KeyboardCowboys.util.Resources;

@RunWith(Arquillian.class)
public class EmployeeRegistrationTest {
	@Deployment
	public static Archive<?> createTestArchive() {
        File[] libs = Maven.resolver().loadPomFromFile("pom.xml")
                .importRuntimeDependencies().as(File.class);
        WebArchive res = ShrinkWrap
				.create(WebArchive.class, "EmployeeRegTest.war")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.controller")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.converter")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.data")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.exceptions")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.interfaces")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.model")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.qualifiers")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.service")
                .addPackage("bcit.ca.infosys.KeyboardCowboys.util")
				.addAsResource("META-INF/test-persistence.xml",
						"META-INF/persistence.xml")
				.addAsWebInfResource(EmptyAsset.INSTANCE, "beans.xml")
				// Deploy our test datasource
				.addAsWebInfResource("test-ds.xml", "test-ds.xml");
        for (File file : libs) {
            res.addAsLibrary(file);
        }
        return res;
	}

	@Inject
	EmployeeRegistrationInterface employeeRegistration;

	@Inject
	Logger log;

	@Test
	public void testRegister() throws Exception {
		Employee newEmployee = new Employee();
		newEmployee.setEmpFName("Jane");
		newEmployee.setEmpLName("Doe");
		newEmployee.setEmpRole("Taking Care of Business");
		newEmployee.setEmpUserName("username");
		EmployeePayInfo testEmpPayInfo = new EmployeePayInfo();
		newEmployee.setEmpPayInfo(testEmpPayInfo);
		employeeRegistration.register(newEmployee);

		assertNotNull(newEmployee.getEmpID());
		log.info(newEmployee.getEmpLName() + " was persisted with id "
				+ newEmployee.getEmpID());
	}

}
